Refactors sum to use generic reduction templates
