+++ dpctl/tensor/libtensor/source/tensor_py.cpp
@@ -49,7 +49,6 @@
 #include "reduction_over_axis.hpp"
 #include "repeat.hpp"
 #include "simplify_iteration_space.hpp"
-#include "sum_reductions.hpp"
 #include "triul_ctor.hpp"
 #include "utils/memory_overlap.hpp"
 #include "utils/strided_iters.hpp"
@@ -413,6 +412,5 @@ PYBIND11_MODULE(_tensor_impl, m)
 
     dpctl::tensor::py_internal::init_elementwise_functions(m);
     dpctl::tensor::py_internal::init_boolean_reduction_functions(m);
-    dpctl::tensor::py_internal::init_sum_reduction_functions(m);
     dpctl::tensor::py_internal::init_reduction_functions(m);
 }
